test: adapt test case of wait with short gap
The thread with the TimedWait should do a Post after the TimedWait because with the new implementation it is expected to succeed and otherwise the semaphore would be held indefinetly.

Between Posting the initial locked semaphore and doing a Wait again there needs to be a small sleep otherwise other process/thread won't be able to claim the semaphore.

This test still fails because fo the TimedWait implementation.